Split config editor row rendering
